backend/requirements.txt
fastapi
uvicorn
supabase
python-dotenv
pydantic[email]
pydantic-settings